use a GET-param, not a preference
