+++ src/argus/htmx/templates/htmx/incident/_incident_list_refresh_info.html
@@ -61,20 +61,7 @@
     <div class="stat py-2">
       <dt class="stat-title text-neutral-content/80">Timeframe</dt>
       <dd class="stat-value text-base">
-        <select class="select select-xs bg-transparent text-base border-none -ml-2"
-                name="timeframe"
-                hx-post="{% url 'htmx:update-preferences' namespace='argus_htmx' %}"
-                hx-headers='{"X-CSRFToken": "{{ csrf_token }}"}'
-                hx-indicator="#incident-list .htmx-indicator"
-                autocomplete="off">
-          {% for timeframe in preferences_choices.argus_htmx.timeframe %}
-            <option class="bg-neutral"
-                    value="{{ timeframe.0|default:0 }}"
-                    {% if timeframe.0 == preferences.argus_htmx.timeframe %}selected{% endif %}>
-              {{ timeframe.1 }}
-            </option>
-          {% endfor %}
-        </select>
+        {{ timeframe_form.timeframe }}
       </dd>
     </div>
   </dl>
