+++ src/argus/htmx/templates/htmx/user/preferences.html
@@ -14,7 +14,6 @@
           <li>{% include "./_dateformat_dropdown.html" %}</li>
           <li>{% include "./_page_size_dropdown.html" %}</li>
           <li>{% include "./_update_interval_dropdown.html" %}</li>
-          <li>{% include "./_timeframe_dropdown.html" %}</li>
         {% endblock preferences %}
       </ul>
     </section>
